Refactor controllers and add docs
