+++ src/MainMenu/MainMenuController.java
@@ -5,20 +5,15 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
 import Utils.SceneLoader.SceneLoader;
+import Utils.StageAwareController;
 
-public class MainMenuController {
+/**
+ * Startmenü des Vokabeltrainers.
+ */
+public class MainMenuController extends StageAwareController {
 
-     /* Need for Stage change over functions include this code block in every Controller Class
-    Pls do not delete or change
-     */
-
-    private Stage stage = null;
-    public void setStage(Stage stage) {
-        this.stage = stage;
-    }
-    ////////////////////////
+    // Stage wird über die Basisklasse gesetzt
     @FXML
     private Button button;
 
@@ -26,23 +21,36 @@ public class MainMenuController {
     private VBox sideMenu;
 
     @FXML
+    /**
+     * Öffnet den eigentlichen Trainer.
+     */
     public void openTrainer(ActionEvent event) {
         SceneLoader.load("/Trainer/Trainer.fxml");
     }
-    public void openTrainerView(ActionEvent event) {
-        SceneLoader.load("/Trainer/TrainerView.fxml");
-    }
+    /**
+     * Öffnet die Einstellungen.
+     */
     public void openSettings(ActionEvent event) {
         SceneLoader.load("/Settings/Settings.fxml");
     }
+
+    /**
+     * Öffnet die Benutzerverwaltung.
+     */
     public void openUserManagement(ActionEvent event) {
         SceneLoader.load("/UserManagement/UserManagement.fxml");
     }
 
+    /**
+     * Zeigt die Highscore-Tabelle an.
+     */
     public void openScoreBoard(ActionEvent event) {
         SceneLoader.load("/ScoreBoard/ScoreBoard.fxml");
     }
 
+    /**
+     * Blendet das seitliche Menü ein oder aus.
+     */
     @FXML
     private void toggleSideMenu() {
         boolean isVisible = sideMenu.isVisible();
@@ -50,6 +58,9 @@ public class MainMenuController {
         sideMenu.setManaged(!isVisible);
     }
 
+    /**
+     * Beendet die Anwendung.
+     */
     @FXML
     private void handleExit() {
         Platform.exit();
