+++ src/Trainer/TrainerController.java
@@ -12,14 +12,17 @@ import javafx.scene.control.TextField;
 import javafx.geometry.Pos;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
 import Utils.SceneLoader.SceneLoader;
+import Utils.StageAwareController;
 
 
 import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
 
-public class TrainerController {
+/**
+ * Steuert den Ablauf des Trainings und wertet die Antworten aus.
+ */
+public class TrainerController extends StageAwareController {
 
     @FXML
     private VBox vocabBox;
@@ -36,7 +39,6 @@ public class TrainerController {
     private String mode = "Englisch zu Deutsch";
     private int currentIndex = 0;
     private int questionsPerRound = 5;
-    private Stage stage;
 
     private static class VocabEntry {
         String solution;
@@ -44,10 +46,9 @@ public class TrainerController {
         HBox container;
     }
 
-    public void setStage(Stage stage) {
-        this.stage = stage;
-    }
-
+    /**
+     * Initialisiert den Trainer und lädt die ersten Vokabeln.
+     */
     @FXML
     private void initialize() {
         userManager.loadFromFile();
@@ -65,6 +66,9 @@ public class TrainerController {
         backButton.setOnAction(event -> SceneLoader.load(stage, "/MainMenu/mainMenu.fxml"));
     }
 
+    /**
+     * Baut die nächste Liste an Vokabeln entsprechend des gewählten Modus auf.
+     */
     private void loadNextVocabSet() {
         vocabBox.getChildren().clear();
         vocabEntries.clear();
@@ -114,6 +118,9 @@ public class TrainerController {
         }
     }
 
+    /**
+     * Prüft alle Eingaben und zeigt richtige bzw. falsche Buchstaben an.
+     */
     public void checkAnswers() {
         int correctCount = 0;
