+++ src/UserManagement/UserManagementController.java
@@ -2,17 +2,20 @@ package UserManagement;
 
 import Utils.SceneLoader.SceneLoader;
 import Utils.UserScore.UserSystem;
+import Utils.StageAwareController;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
-import javafx.stage.Stage;
 
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class UserManagementController implements Initializable, SceneLoader.HasStage {
+/**
+ * Verwaltung der vorhandenen Benutzer.
+ */
+public class UserManagementController extends StageAwareController implements Initializable {
     @FXML
     private TextField newUserField;
     @FXML
@@ -22,13 +25,7 @@ public class UserManagementController implements Initializable, SceneLoader.HasS
     @FXML
     private Button backButton;
 
-    private Stage stage;
     private final UserSystem userSystem = UserSystem.getInstance();
-
-    @Override
-    public void setStage(Stage stage) {
-        this.stage = stage;
-    }
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
@@ -37,6 +34,9 @@ public class UserManagementController implements Initializable, SceneLoader.HasS
         searchField.textProperty().addListener((obs, o, n) -> refreshList(n));
     }
 
+    /**
+     * Erstellt einen neuen Benutzer und wählt ihn anschließend aus.
+     */
     @FXML
     private void createUser() {
         String name = newUserField.getText().trim();
@@ -49,6 +49,9 @@ public class UserManagementController implements Initializable, SceneLoader.HasS
         }
     }
 
+    /**
+     * Setzt den aktuell ausgewählten Benutzer.
+     */
     @FXML
     private void selectUser() {
         String selected = userList.getSelectionModel().getSelectedItem();
@@ -58,6 +61,9 @@ public class UserManagementController implements Initializable, SceneLoader.HasS
         }
     }
 
+    /**
+     * Aktualisiert die Liste der Benutzer unter Berücksichtigung eines Filters.
+     */
     private void refreshList(String filter) {
         userList.getItems().clear();
         for (String name : userSystem.getAllUserNames()) {
@@ -67,6 +73,9 @@ public class UserManagementController implements Initializable, SceneLoader.HasS
         }
     }
 
+    /**
+     * Zurück zum Hauptmenü.
+     */
     @FXML
     private void backToMenu() {
         SceneLoader.load(stage, "/MainMenu/mainMenu.fxml");
